redid migration

+++ Garage2/Migrations/Configuration.cs
@@ -35,33 +35,37 @@ namespace Garage2.Migrations
                 new Owner { Name = "Jane Doe" }
                 );
 
-            context.Vehicles.AddOrUpdate(v => v.Reg,
-                new Vehicle { OwnerId = 1, Reg = "AAA111", OwnerInfo = null, VehicleTypeId = 1 },
-                new Vehicle { OwnerId = 1, Reg = "BBB222", OwnerInfo = null, VehicleTypeId = 2 },
-                new Vehicle { OwnerId = 2, Reg = "CCC333", OwnerInfo = null, VehicleTypeId = 3 },
-                new Vehicle { OwnerId = 3, Reg = "DDD444", OwnerInfo = null, VehicleTypeId = 4 }
-                );
+            if (context.VehicleTypes.Count() > 0)
+            {
 
-            if (context.Parkings.Count() == 0)
-            {
-                context.Parkings.AddOrUpdate(
-                    new Parking { ParkingSlotId = 1, VehicleReg = "AAA111", DateIn = new DateTime(2016, 2, 15), DateOut = null },
-                    new Parking { ParkingSlotId = 2, VehicleReg = "BBB222", DateIn = new DateTime(2016, 2, 16), DateOut = null },
-                    new Parking { ParkingSlotId = 3, VehicleReg = "CCC333", DateIn = new DateTime(2016, 2, 17), DateOut = null },
-                    new Parking { ParkingSlotId = 4, VehicleReg = "DDD444", DateIn = new DateTime(2016, 2, 17), DateOut = new DateTime(2016, 2, 18) },
-                    new Parking { ParkingSlotId = 5, VehicleReg = "DDD444", DateIn = new DateTime(2016, 2, 18), DateOut = null }
+                context.Vehicles.AddOrUpdate(v => v.Reg,
+                    new Vehicle { OwnerId = 1, Reg = "AAA111", OwnerInfo = null, VehicleTypeId = 1 },
+                    new Vehicle { OwnerId = 1, Reg = "BBB222", OwnerInfo = null, VehicleTypeId = 2 },
+                    new Vehicle { OwnerId = 2, Reg = "CCC333", OwnerInfo = null, VehicleTypeId = 3 },
+                    new Vehicle { OwnerId = 3, Reg = "DDD444", OwnerInfo = null, VehicleTypeId = 4 }
                     );
-                context.ParkingSlots.AddOrUpdate(s => s.Id,
-                    new ParkingSlot { Id = 1, Occupied = true, VehicleReg = "AAA111" },
-                    new ParkingSlot { Id = 2, Occupied = true, VehicleReg = "BBB222" },
-                    new ParkingSlot { Id = 3, Occupied = true, VehicleReg = "CCC333" },
-                    new ParkingSlot { Id = 5, Occupied = true, VehicleReg = "DDD444" }
-                    );
+
+                if (context.Parkings.Count() == 0)
+                {
+                    context.Parkings.AddOrUpdate(
+                        new Parking { ParkingSlotId = 1, VehicleReg = "AAA111", DateIn = new DateTime(2016, 2, 15), DateOut = null },
+                        new Parking { ParkingSlotId = 2, VehicleReg = "BBB222", DateIn = new DateTime(2016, 2, 16), DateOut = null },
+                        new Parking { ParkingSlotId = 3, VehicleReg = "CCC333", DateIn = new DateTime(2016, 2, 17), DateOut = null },
+                        new Parking { ParkingSlotId = 4, VehicleReg = "DDD444", DateIn = new DateTime(2016, 2, 17), DateOut = new DateTime(2016, 2, 18) },
+                        new Parking { ParkingSlotId = 5, VehicleReg = "DDD444", DateIn = new DateTime(2016, 2, 18), DateOut = null }
+                        );
+                    context.ParkingSlots.AddOrUpdate(s => s.Id,
+                        new ParkingSlot { Id = 1, Occupied = true, VehicleReg = "AAA111" },
+                        new ParkingSlot { Id = 2, Occupied = true, VehicleReg = "BBB222" },
+                        new ParkingSlot { Id = 3, Occupied = true, VehicleReg = "CCC333" },
+                        new ParkingSlot { Id = 5, Occupied = true, VehicleReg = "DDD444" }
+                        );
+                }
+                /*else
+                {
+                    context.Parkings.RemoveRange(context.Parkings);
+                }*/
             }
-            /*else
-            {
-                context.Parkings.RemoveRange(context.Parkings);
-            }*/
 
             context.SaveChanges();
         }
